[CHORE] #50 : 라벨 호출 이미지 변경 recipe -> products
