+++ client/app/src/main/java/com/example/healthc/data/service/ProductService.kt
@@ -3,6 +3,7 @@ package com.example.healthc.data.service
 import com.example.healthc.BuildConfig
 import com.example.healthc.data.model.remote.product.ProductItemResultResponse
 import com.example.healthc.data.model.remote.product.ProductResponse
+import okhttp3.ResponseBody
 import retrofit2.http.*
 
 interface ProductService {
@@ -20,4 +21,12 @@ interface ProductService {
         @Query("query") query : String,
         @Query("number") number : Int = 10
     ) : ProductItemResultResponse
+
+    // 가공식품 라벨 찾기
+    @GET("/food/products/{id}/nutritionLabel.png")
+    suspend fun getNutritionLabel(
+        @Path("id") id : Int,
+        @Query("apiKey") apiKey: String = BuildConfig.SPOON_API_KEY,
+        @Query("showIngredients") showIngredient : Boolean = true
+    ) : ResponseBody
 }
